requirements.txt
streamlit
pandas
numpy
scikit-learn
yfinance
fredapi
pandas-ta
matplotlib
python-dotenv
